Moved renderer and main files to their own directory
