+++ renderer/index.html
@@ -3,9 +3,9 @@
 	<head>
 		<meta charset="UTF-8">
 		<title>Hello World!</title>
-		<link href="node_modules/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet" type="text/css" media="all" />
-		<link href="node_modules/angular-ivh-treeview/dist/ivh-treeview.css" rel="stylesheet" type="text/css" media="all" />
-		<link href="node_modules/angular-ivh-treeview/dist/ivh-treeview-theme-basic.css" rel="stylesheet" type="text/css" media="all" />
+		<link href="../node_modules/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet" type="text/css" media="all" />
+		<link href="../node_modules/angular-ivh-treeview/dist/ivh-treeview.css" rel="stylesheet" type="text/css" media="all" />
+		<link href="../node_modules/angular-ivh-treeview/dist/ivh-treeview-theme-basic.css" rel="stylesheet" type="text/css" media="all" />
 
 		<link href="assets/css/main-ui.css" rel="stylesheet" type="text/css" />
 		<link href="assets/icons/atlas.svg" type="image/svg+xml" />
